Issue #535: Moves the bottom container into a component.
